Split tmux session drawer
